breakout pump subscriber for easier testing
